+++ assets/src/adaptivity/operators/contains.ts
@@ -5,20 +5,44 @@ export const containsOperator = (inputValue: any, conditionValue: any) => {
     return false;
   }
 
+  // always read as: Does the INPUT contain the CONDITION?
+  /* console.log('containsOperator', { inputValue, conditionValue }); */
+
   if (looksLikeAnArray(conditionValue)) {
     const conditionArray = parseArray(conditionValue);
     if (looksLikeAnArray(inputValue)) {
       const inputArray = parseArray(inputValue);
       // if the input is an array, the condition array should contain every one of the input array values
-      return inputArray.every((item) => conditionArray.includes(item));
+      // does [1, 2, 3] contain [1, 3]?
+      return conditionArray.every((item) => inputArray.includes(item));
     } else {
-      return conditionArray.includes(inputValue);
+      // does 'abc' contain ['a', 'b']? (contains both, case insensitive)
+      return conditionArray.every((item) => {
+        if (isString(item)) {
+          return inputValue.toLocaleLowerCase().includes((item as string).toLocaleLowerCase());
+        } else {
+          return inputValue.includes(item);
+        }
+      });
     }
   }
 
-  if (isString(conditionValue)) {
-    if (isString(inputValue)) {
-      return conditionValue.toLocaleLowerCase().includes(inputValue.toLocaleLowerCase());
+  if (looksLikeAnArray(inputValue)) {
+    const inputArray = parseArray(inputValue);
+    // does ['a', 'b'] contain 'A'? (contains, case insensitive)
+    return inputArray.some((item) => {
+      if (isString(item)) {
+        return (item as string).toLocaleLowerCase().includes(conditionValue.toLocaleLowerCase());
+      }
+      return item === conditionValue;
+    });
+  }
+
+  if (isString(inputValue)) {
+    if (isString(conditionValue)) {
+      return inputValue.toLocaleLowerCase().includes(conditionValue.toLocaleLowerCase());
+    } else {
+      return inputValue.includes(conditionValue);
     }
   }
 
@@ -64,24 +88,47 @@ export const containsOnlyOperator = (inputValue: any, conditionValue: any) => {
   return updatedFacts.every((fact: any) => updatedValues.includes(fact));
 };
 
+// case sensitive version of containsOperator
 export const containsExactlyOperator = (inputValue: any, conditionValue: any) => {
   // inputValue is exactly equal to conditionValue
   if (!conditionValue || !inputValue) {
     return false;
   }
 
-  // We are parseNumString for the cases where inputValue contains numbers but the values contain strings or vice-versa
-  const updatedFacts = parseArray(inputValue);
-  const updatedValues = parseArray(conditionValue);
+  // always read as: Does the INPUT contain the CONDITION?
+  /* console.log('containsExactlyOperator', { inputValue, conditionValue }); */
 
-  if (Array.isArray(inputValue) && Array.isArray(conditionValue)) {
-    return (
-      updatedFacts.every((fact) => updatedValues.includes(fact)) &&
-      updatedFacts.length == updatedValues.length
-    );
-  } else {
-    return inputValue === conditionValue;
+  if (looksLikeAnArray(conditionValue)) {
+    const conditionArray = parseArray(conditionValue);
+    if (looksLikeAnArray(inputValue)) {
+      const inputArray = parseArray(inputValue);
+      // if the input is an array, the condition array should contain every one of the input array values
+      // does [1, 2, 3] contain [1, 3]?
+      return conditionArray.every((item) => inputArray.includes(item));
+    } else {
+      // does 'abc' contain ['a', 'b']? (contains both, case sensitive)
+      return conditionArray.every((item) => {
+        return inputValue.includes(item);
+      });
+    }
   }
+
+  if (looksLikeAnArray(inputValue)) {
+    const inputArray = parseArray(inputValue);
+    // does ['a', 'b'] contain 'A'? (contains, case insensitive)
+    return inputArray.some((item) => {
+      if (isString(item)) {
+        return (item as string).includes(conditionValue);
+      }
+      return item === conditionValue;
+    });
+  }
+
+  if (isString(inputValue)) {
+    return inputValue.includes(conditionValue);
+  }
+
+  return false;
 };
 
 export const notContainsExactlyOperator = (inputValue: any, conditionValue: any) =>