+++ lib/oli/accounts.ex
@@ -2,6 +2,8 @@ defmodule Oli.Accounts do
   import Ecto.Query, warn: false
   import Oli.Utils, only: [value_or: 2]
 
+  alias Ecto.Multi
+
   alias Oli.Accounts.{
     User,
     Author,
@@ -11,6 +13,7 @@ defmodule Oli.Accounts do
     AuthorPreferences
   }
 
+  alias Oli.Groups
   alias Oli.Groups.CommunityAccount
   alias Oli.Repo
   alias Oli.Repo.{Paging, Sorting}
@@ -629,5 +632,44 @@ defmodule Oli.Accounts do
   def user_confirmation_pending?(user) do
     EmailConfirmationContext.current_email_unconfirmed?(user, []) or
       EmailConfirmationContext.pending_email_change?(user, [])
+  end
+
+  @doc """
+  Inserts or updates an user logged in via sso, and adds the user as a member of the given community.
+
+  ## Examples
+
+      iex> setup_sso_user(fields, community_id)
+      {:ok, %User{}}    -> # Inserted or updated with success
+      {:error, changeset}         -> # Something went wrong
+
+  """
+  def setup_sso_user(fields, community_id) do
+    res =
+      Multi.new()
+      |> Multi.run(:user, &create_sso_user(&1, &2, fields))
+      |> Multi.run(:community_account, &create_community_account(&1, &2, community_id))
+      |> Repo.transaction()
+
+    case res do
+      {:ok, %{user: user}} ->
+        {:ok, user}
+
+      {:error, _, changeset, _} ->
+        {:error, changeset}
+    end
+  end
+
+  defp create_sso_user(_repo, _changes, fields) do
+    insert_or_update_lms_user(%{
+      sub: Map.get(fields, "sub"),
+      preferred_username: Map.get(fields, "cognito:username"),
+      email: Map.get(fields, "email"),
+      can_create_sections: true
+    })
+  end
+
+  defp create_community_account(_repo, %{user: %User{id: user_id}}, community_id) do
+    Groups.find_or_create_community_user_account(user_id, community_id)
   end
 end