+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Instructor review of completed graded attempts
 - Allow gates to be defined in products
 - Hide subsequent purpose types for activities when the same purpose type is used in a series
+- Allow students to pay and apply codes during a grace period
 
 ## 0.18.4 (2022-02-24)